Fixing the Remove robot so that it also deletes a talk page, if one exists.
The default behavior of the Removal robot has been altered so that it uses
a custom edit summary (if one is given) in preference to the default edit
summary when performing a deletion of the category page.  This behavior is
most likely preferred because a custom edit summary is almost guaranteed to
contain information that is more helpful than "Category was disbanded".