Added support for importing pywikibot instead of wikipedia.
The used method is not very nice, but quite effective. Because wikipedia.py has
many globals and dependencies, it's not easy to create a correct import chain.
Therefore, instead of importing wikipedia in the pywikibot module, wikipedia.py
monkey-patches /itself/ in there.

Long story short: import pywikibot and be done with it.